Fix 500 error when searching for one tag.
This fixes by making the search an OR search for that tag, as Jocelyn
advises me this would be a preferable behaviour change to the pervious
version, which did the AND search (i.e. things exactly tagged).

For paranoia's sake, I also fix the code so that if we're sent an operator
we don't understand, then it's just entirely ignored (meaning filtering
on tags isn't done).